requirements.txt
Django==2.2.28
pytz==2021.3
sqlparse==0.4.2
requests==2.26.0
Pillow==9.3.0
sorl-thumbnail==12.7.0
django-cleanup==6.0.0
django-simple-captcha==0.5.17
python-dotenv==0.21.0
python-dateutil==2.8.2